Significant naming changes and fixes
*Breaking Change*
Many fields within generated code had poor naming standards, such as the
word `Native` being capitalized as `NATive` and so on. Numerous
abbreviated fields also made these field names close to unreadable. This
is an attempt to fix many of these issues so the API is much more user
friendly.

This change corrects many of them, but does so as a breaking change to
the API. This may require a larger version bump to ensure others are
aware.